requirements.txt
flask==2.3.3
flask_login==0.6.2
flask_sqlalchemy==3.1.1
flask_migrate==4.0.5
sqlalchemy==2.0.29
pillow==9.5.0
werkzeug==2.3.7
ultralytics==8.0.207
python-dotenv==1.0.0
gunicorn==21.2.0
opencv-python==4.9.0.80
matplotlib==3.8.2
numpy==1.26.2
pyyaml==6.0.1
tqdm==4.66.1
psycopg2-binary==2.9.9
supervision==0.19.0 